refactor(tests): split template compiler tests intop smaller files
