+++ packages/__tests__/3-runtime-html/template-compiler.spec.ts
@@ -2,13 +2,7 @@ import {
   Constructable,
   IContainer,
   kebabCase,
-  ISink,
-  ILogEvent,
-  LoggerConfiguration,
-  DefaultLogger,
-  LogLevel,
   camelCase,
-  Registration,
 } from '@aurelia/kernel';
 import {
   ExpressionType,
@@ -34,17 +28,12 @@ import {
   PartialCustomAttributeDefinition,
   CustomElementDefinition,
   IInstruction,
-  CustomElementType,
-  AuSlot,
-  Aurelia,
   HydrateElementInstruction,
   HydrateTemplateController,
   InstructionType as HTT,
   InstructionType as TT,
   HydrateLetElementInstruction,
-  InstructionType,
   CustomAttributeDefinition,
-  ITemplateCompilerHooks,
   IAttributeParser,
   HydrateAttributeInstruction,
   AttrSyntax,
@@ -58,7 +47,6 @@ import {
   eachCartesianJoinFactory,
   TestContext,
   verifyBindingInstructionsEqual,
-  generateCartesianProduct,
 } from '@aurelia/testing';
 
 export function createAttribute(name: string, value: string): Attr {
@@ -67,7 +55,7 @@ export function createAttribute(name: string, value: string): Attr {
   return attr;
 }
 
-describe('template-compiler.spec.ts\n  [TemplateCompiler]', function () {
+describe('3-runtime-html/template-compiler.spec.ts', function () {
   let ctx: TestContext;
   let sut: ITemplateCompiler;
   let container: IContainer;
@@ -1860,796 +1848,6 @@ describe(`TemplateCompiler - combinations`, function () {
         ])]
       );
     });
-  });
-});
-
-describe('TemplateCompiler - local templates', function () {
-  class EventLog implements ISink {
-    public readonly log: ILogEvent[] = [];
-    public handleEvent(event: ILogEvent): void {
-      this.log.push(event);
-    }
-  }
-  function createFixture() {
-    const ctx = TestContext.create();
-    const container = ctx.container;
-    container.register(LoggerConfiguration.create({ sinks: [EventLog] }));
-    const sut = ctx.templateCompiler;
-    return { ctx, container, sut };
-  }
-
-  class LocalTemplateTestData {
-
-    public constructor(
-      public readonly template: string,
-      private readonly expectedResources: Map<string, ElementInfo>,
-      private readonly templateFreq: Map<string, number>,
-      public readonly expectedContent: string,
-    ) {
-      this.verifyDefinition = this.verifyDefinition.bind(this);
-    }
-    public verifyDefinition(definition: CustomElementDefinition, container: IContainer): void {
-      assert.equal((definition.template as HTMLTemplateElement).querySelector('template[as-custom-element]'), null);
-
-      for (const [name, info] of this.expectedResources) {
-        assert.deepStrictEqual(ElementInfo.from(container.find(CustomElement, name), void 0), info, 'element info');
-      }
-      const ceInstructions: HydrateElementInstruction[] = definition.instructions.flatMap((i) => i).filter((i) => i instanceof HydrateElementInstruction) as HydrateElementInstruction[];
-      for (const [template, freq] of this.templateFreq) {
-        assert.strictEqual(ceInstructions.filter((i) => i.res === template).length, freq, 'HydrateElementInstruction.freq');
-      }
-    }
-
-  }
-  function *getLocalTemplateTestData() {
-    yield new LocalTemplateTestData(
-      `<template as-custom-element="foo-bar">static</template>
-      <foo-bar></foo-bar>`,
-      new Map([['foo-bar', new ElementInfo('foo-bar', void 0, false)]]),
-      new Map([['foo-bar', 1]]),
-      'static'
-    );
-    yield new LocalTemplateTestData(
-      `<foo-bar></foo-bar>
-      <template as-custom-element="foo-bar">static</template>`,
-      new Map([['foo-bar', new ElementInfo('foo-bar', void 0, false)]]),
-      new Map([['foo-bar', 1]]),
-      'static'
-    );
-    yield new LocalTemplateTestData(
-      `<foo-bar></foo-bar>
-      <template as-custom-element="foo-bar">static</template>
-      <foo-bar></foo-bar>`,
-      new Map([['foo-bar', new ElementInfo('foo-bar', void 0, false)]]),
-      new Map([['foo-bar', 2]]),
-      'static static'
-    );
-    yield new LocalTemplateTestData(
-      `<template as-custom-element="foo-bar">static foo-bar</template>
-      <template as-custom-element="fiz-baz">static fiz-baz</template>
-      <fiz-baz></fiz-baz>
-      <foo-bar></foo-bar>`,
-      new Map([['foo-bar', new ElementInfo('foo-bar', void 0, false)], ['fiz-baz', new ElementInfo('fiz-baz', void 0, false)]]),
-      new Map([['foo-bar', 1], ['fiz-baz', 1]]),
-      'static fiz-baz static foo-bar'
-    );
-    const bindingModeMap = new Map([
-      ['oneTime', BindingMode.oneTime],
-      ['toView', BindingMode.toView],
-      ['fromView', BindingMode.fromView],
-      ['twoWay', BindingMode.twoWay],
-      ['default', BindingMode.toView],
-    ]);
-    for (const [bindingMode, props, attributeName] of generateCartesianProduct([
-      [...bindingModeMap.keys(), void 0],
-      [['prop'], ['prop', 'camelProp']],
-      ['fiz-baz', undefined],
-    ])) {
-      const ei = new ElementInfo('foo-bar', void 0, false);
-      const mode = bindingModeMap.get(bindingMode) ?? BindingMode.toView;
-      let bindables = '';
-      let templateBody = '';
-      let attrExpr = '';
-      let renderedContent = '';
-      const value = "awesome possum";
-      for (let i = 0, ii = props.length; i < ii; i++) {
-        const prop = props[i];
-        const bi = new BindableInfo(prop, mode);
-        const attr = kebabCase(attributeName !== void 0 ? `${attributeName}${i + 1}` : prop);
-        ei.bindables[attr] = bi;
-
-        bindables += `<bindable property='${prop}'${bindingMode !== void 0 ? ` mode="${bindingMode}"` : ''}${attributeName !== void 0 ? ` attribute="${attr}"` : ''}></bindable>`;
-        templateBody += ` \${${prop}}`;
-        const content = `${value}${i + 1}`;
-        attrExpr += ` ${attr}="${content}"`;
-        renderedContent += ` ${content}`;
-      }
-
-      yield new LocalTemplateTestData(
-        `<template as-custom-element="foo-bar">
-            ${bindables}
-            ${templateBody}
-          </template>
-          <foo-bar ${attrExpr}></foo-bar>`,
-        new Map([['foo-bar', ei]]),
-        new Map([['foo-bar', 1]]),
-        renderedContent.trim()
-      );
-    }
-  }
-  for (const { template, verifyDefinition, expectedContent } of getLocalTemplateTestData()) {
-    it(template, function () {
-      const { container, sut } = createFixture();
-      sut.resolveResources = false;
-      const definition = sut.compile({ name: 'lorem-ipsum', template }, container, null);
-      verifyDefinition(definition, container);
-    });
-    if (template.includes(`mode="fromView"`)) { continue; }
-    it(`${template} - content`, async function () {
-      const { ctx, container } = createFixture();
-      const host = ctx.doc.createElement('div');
-      ctx.doc.body.appendChild(host);
-      const au = new Aurelia(container)
-        .app({ host, component: CustomElement.define({ name: 'lorem-ipsum', template }, class { }) });
-
-      await au.start();
-
-      assert.html.textContent(host, expectedContent);
-
-      await au.stop();
-      ctx.doc.body.removeChild(host);
-      au.dispose();
-    });
-  }
-
-  it('works with if', async function () {
-    const template = `<template as-custom-element="foo-bar">
-    <bindable property='prop'></bindable>
-    \${prop}
-   </template>
-   <foo-bar prop="awesome possum" if.bind="true"></foo-bar>
-   <foo-bar prop="ignored" if.bind="false"></foo-bar>`;
-    const expectedContent = "awesome possum";
-
-    const { ctx, container } = createFixture();
-    const host = ctx.doc.createElement('div');
-    ctx.doc.body.appendChild(host);
-    const au = new Aurelia(container)
-      .app({ host, component: CustomElement.define({ name: 'lorem-ipsum', template }, class { }) });
-
-    await au.start();
-
-    assert.html.textContent(host, expectedContent);
-
-    await au.stop();
-    ctx.doc.body.removeChild(host);
-    au.dispose();
-  });
-
-  it('works with for', async function () {
-    const template = `<template as-custom-element="foo-bar">
-    <bindable property='prop'></bindable>
-    \${prop}
-   </template>
-   <foo-bar repeat.for="i of 5" prop.bind="i"></foo-bar>`;
-    const expectedContent = "0 1 2 3 4";
-
-    const { ctx, container } = createFixture();
-    const host = ctx.doc.createElement('div');
-    ctx.doc.body.appendChild(host);
-    const au = new Aurelia(container)
-      .app({ host, component: CustomElement.define({ name: 'lorem-ipsum', template }, class { }) });
-
-    await au.start();
-
-    assert.html.textContent(host, expectedContent);
-
-    await au.stop();
-    ctx.doc.body.removeChild(host);
-    au.dispose();
-  });
-
-  it('works with nested templates - 1', async function () {
-    @customElement({ name: 'level-one', template: `<template as-custom-element="foo-bar"><bindable property='prop'></bindable>Level One \${prop}</template><foo-bar prop.bind="prop"></foo-bar>` })
-    class LevelOne {
-      @bindable public prop: string;
-    }
-    @customElement({
-      name: 'level-two', template: `
-    <template as-custom-element="foo-bar">
-      <bindable property='prop'></bindable>
-      Level Two \${prop}
-      <level-one prop="inter-dimensional portal"></level-one>
-    </template>
-    <foo-bar prop.bind="prop"></foo-bar>
-    <level-one prop.bind="prop"></level-one>
-    `})
-    class LevelTwo {
-      @bindable public prop: string;
-    }
-    const template = `<level-two prop="foo2"></level-two><level-one prop="foo1"></level-one>`;
-    const expectedContent = "Level Two foo2 Level One inter-dimensional portal Level One foo2 Level One foo1";
-
-    const { ctx, container } = createFixture();
-    const host = ctx.doc.createElement('div');
-    ctx.doc.body.appendChild(host);
-    const au = new Aurelia(container)
-      .register(LevelOne, LevelTwo)
-      .app({ host, component: CustomElement.define({ name: 'lorem-ipsum', template }, class { }) });
-
-    await au.start();
-
-    assert.html.textContent(host, expectedContent);
-
-    await au.stop();
-    ctx.doc.body.removeChild(host);
-    au.dispose();
-  });
-
-  it('works with nested templates - 2', async function () {
-    const template = `
-    <template as-custom-element="el-one">
-      <template as-custom-element="one-two">
-        1
-      </template>
-      2
-      <one-two></one-two>
-    </template>
-    <template as-custom-element="el-two">
-      <template as-custom-element="two-two">
-        3
-      </template>
-      4
-      <two-two></two-two>
-    </template>
-    <el-two></el-two>
-    <el-one></el-one>
-    `;
-    const expectedContent = "4 3 2 1";
-
-    const { ctx, container } = createFixture();
-    const host = ctx.doc.createElement('div');
-    ctx.doc.body.appendChild(host);
-    const au = new Aurelia(container)
-      .app({ host, component: CustomElement.define({ name: 'lorem-ipsum', template }, class { }) });
-
-    await au.start();
-
-    assert.html.textContent(host, expectedContent);
-
-    await au.stop();
-    ctx.doc.body.removeChild(host);
-    au.dispose();
-  });
-
-  it('works with non-global dependencies in owning template', async function () {
-    @customElement({ name: 'my-ce', template: 'my-ce-content' })
-    class MyCe { }
-
-    const template = `
-    <my-ce></my-ce>
-    <my-le></my-le>
-    <template as-custom-element="my-le">
-      my-le-content
-      <my-ce></my-ce>
-    </template>
-    `;
-    @customElement({ name: 'my-app', template, dependencies: [MyCe] })
-    class App { }
-
-    const { ctx, container } = createFixture();
-    const host = ctx.doc.createElement('div');
-    ctx.doc.body.appendChild(host);
-    const au = new Aurelia(container)
-      .app({ host, component: App });
-
-    await au.start();
-
-    assert.html.textContent(host, 'my-ce-content my-le-content my-ce-content');
-
-    await au.stop();
-    ctx.doc.body.removeChild(host);
-    au.dispose();
-  });
-
-  it('works with non-global dependencies - template-controllers - if', async function () {
-    @customElement({ name: 'my-ce', template: 'my-ce-content' })
-    class MyCe { }
-
-    const template = `
-    <my-ce></my-ce>
-    <my-le if.bind="true"></my-le>
-    <template as-custom-element="my-le">
-      my-le-content
-      <my-ce></my-ce>
-    </template>
-    `;
-    @customElement({ name: 'my-app', template, dependencies: [MyCe] })
-    class App { }
-
-    const { ctx, container } = createFixture();
-    const host = ctx.doc.createElement('div');
-    ctx.doc.body.appendChild(host);
-    const au = new Aurelia(container)
-      .app({ host, component: App });
-
-    await au.start();
-
-    assert.html.textContent(host, 'my-ce-content my-le-content my-ce-content');
-
-    await au.stop();
-    ctx.doc.body.removeChild(host);
-    au.dispose();
-  });
-
-  it('works with non-global dependencies - nested-template-controllers - [repeat.for]>[if]', async function () {
-    @customElement({ name: 'my-ce', template: 'my-ce-content' })
-    class MyCe { }
-
-    const template = `
-    <my-ce></my-ce>
-    <my-le repeat.for="prop of 5" if.bind="prop % 2 === 0" prop.bind></my-le>
-    <template as-custom-element="my-le">
-      <bindable property="prop"></bindable>
-      \${prop}
-      <my-ce></my-ce>
-    </template>
-    `;
-    @customElement({ name: 'my-app', template, dependencies: [MyCe] })
-    class App { }
-
-    const { ctx, container } = createFixture();
-    const host = ctx.doc.createElement('div');
-    ctx.doc.body.appendChild(host);
-    const au = new Aurelia(container)
-      .app({ host, component: App });
-
-    await au.start();
-
-    assert.html.textContent(host, 'my-ce-content 0 my-ce-content 2 my-ce-content 4 my-ce-content');
-
-    await au.stop();
-    ctx.doc.body.removeChild(host);
-    au.dispose();
-  });
-
-  it('recognizes owning element', async function () {
-    const template = `
-      my-app-content
-      <my-le prop.bind></my-le>
-      <template as-custom-element="my-le">
-        <bindable property="prop"></bindable>
-        my-le-content
-        <my-app if.bind="prop"></my-app>
-      </template>`;
-    @customElement({ name: 'my-app', template })
-    class App {
-      public prop = false;
-    }
-
-    const { ctx, container } = createFixture();
-    const host = ctx.doc.createElement('div');
-    ctx.doc.body.appendChild(host);
-    const au = new Aurelia(container)
-      .app({ host, component: App });
-
-    await au.start();
-
-    const vm = au.root.controller.viewModel as App;
-
-    assert.html.textContent(host, 'my-app-content my-le-content');
-
-    vm.prop = true;
-    ctx.platform.domWriteQueue.flush();
-
-    assert.html.textContent(host, 'my-app-content my-le-content my-app-content my-le-content');
-
-    await au.stop();
-    ctx.doc.body.removeChild(host);
-    au.dispose();
-  });
-
-  it('all local elements recognize each other', async function () {
-    const template = `
-      my-app-content
-      <my-le-1></my-le-1>
-      <my-le-2></my-le-2>
-
-      <template as-custom-element="my-le-1">
-        my-le-1-content
-        <my-le-2></my-le-2>
-      </template>
-
-      <template as-custom-element="my-le-2">
-        my-le-2-content
-        <my-le-3></my-le-3>
-      </template>
-
-      <template as-custom-element="my-le-3">
-        my-le-3-content
-      </template>`;
-
-    @customElement({ name: 'my-app', template })
-    class App {}
-
-    const { ctx, container } = createFixture();
-    const host = ctx.doc.createElement('div');
-    ctx.doc.body.appendChild(host);
-    const au = new Aurelia(container)
-      .app({ host, component: App });
-
-    await au.start();
-
-    assert.html.textContent(
-      host,
-      'my-app-content ' +
-      'my-le-1-content my-le-2-content my-le-3-content ' +
-      'my-le-2-content my-le-3-content');
-
-    await au.stop();
-    ctx.doc.body.removeChild(host);
-    au.dispose();
-  });
-
-  it('throws error if a root template is a local template', function () {
-    const template = `<template as-custom-element="foo-bar">I have local root!</template>`;
-    const { container, sut } = createFixture();
-    assert.throws(() => sut.compile({ name: 'lorem-ipsum', template }, container, null), 'The root cannot be a local template itself.');
-  });
-
-  it('throws error if the custom element has only local templates', function () {
-    const template = `
-    <template as-custom-element="foo-bar">Does this work?</template>
-    <template as-custom-element="fiz-baz">Of course not!</template>
-    `;
-    const { container, sut } = createFixture();
-    assert.throws(() => sut.compile({ name: 'lorem-ipsum', template }, container, null), 'The custom element does not have any content other than local template(s).');
-  });
-
-  it('throws error if a local template is not under root', function () {
-    const template = `<div><template as-custom-element="foo-bar">Can I hide here?</template></div>`;
-    const { container, sut } = createFixture();
-    assert.throws(() => sut.compile({ name: 'lorem-ipsum', template }, container, null), 'Local templates needs to be defined directly under root.');
-  });
-
-  it('throws error if a local template does not have name', function () {
-    const template = `<template as-custom-element="">foo-bar</template><div></div>`;
-    const { container, sut } = createFixture();
-    assert.throws(() => sut.compile({ name: 'lorem-ipsum', template }, container, null), 'The value of "as-custom-element" attribute cannot be empty for local template');
-  });
-
-  it('throws error if a duplicate local templates are found', function () {
-    const template = `<template as-custom-element="foo-bar">foo-bar1</template><template as-custom-element="foo-bar">foo-bar2</template><div></div>`;
-    const { container, sut } = createFixture();
-    assert.throws(() => sut.compile({ name: 'lorem-ipsum', template }, container, null), 'Duplicate definition of the local template named foo-bar');
-  });
-
-  it('throws error if bindable is not under root', function () {
-    const template = `<template as-custom-element="foo-bar">
-      <div>
-        <bindable property="prop"></bindable>
-      </div>
-    </template>
-    <div></div>`;
-    const { container, sut } = createFixture();
-    assert.throws(() => sut.compile({ name: 'lorem-ipsum', template }, container, null), 'Bindable properties of local templates needs to be defined directly under root.');
-  });
-
-  it('throws error if bindable property is missing', function () {
-    const template = `<template as-custom-element="foo-bar">
-      <bindable attribute="prop"></bindable>
-    </template>
-    <div></div>`;
-    const { container, sut } = createFixture();
-    assert.throws(() => sut.compile({ name: 'lorem-ipsum', template }, container, null), 'The attribute \'property\' is missing in <bindable attribute="prop"></bindable>');
-  });
-
-  it('throws error if duplicate bindable properties are found', function () {
-    const template = `<template as-custom-element="foo-bar">
-      <bindable property="prop" attribute="bar"></bindable>
-      <bindable property="prop" attribute="baz"></bindable>
-    </template>
-    <div></div>`;
-    const { container, sut } = createFixture();
-    assert.throws(
-      () => sut.compile({ name: 'lorem-ipsum', template }, container, null),
-      'Bindable property and attribute needs to be unique; found property: prop, attribute: '
-    );
-  });
-
-  it('throws error if duplicate bindable attributes are found', function () {
-    const template = `<template as-custom-element="foo-bar">
-      <bindable property="prop1" attribute="bar"></bindable>
-      <bindable property="prop2" attribute="bar"></bindable>
-    </template>
-    <div></div>`;
-    const { container, sut } = createFixture();
-    assert.throws(
-      () => sut.compile({ name: 'lorem-ipsum', template }, container, null),
-      'Bindable property and attribute needs to be unique; found property: prop2, attribute: bar'
-    );
-  });
-
-  for (const attr of ['if.bind="true"', 'if.bind="false"', 'else', 'repeat.for="item of items"', 'with.bind="{a:1}"', 'switch.bind="cond"', 'case="case1"']) {
-    it(`throws error if local-template surrogate has template controller - ${attr}`, function () {
-      const template = `<template as-custom-element="foo-bar" ${attr}>
-      <bindable property="prop1" attribute="bar"></bindable>
-    </template>
-    <foo-bar></foo-bar>`;
-      const { ctx, container } = createFixture();
-
-      assert.throws(() =>
-        new Aurelia(container)
-          .app({ host: ctx.doc.createElement('div'), component: CustomElement.define({ name: 'lorem-ipsum', template }, class { }) }),
-        `Template controller ${attr.split('.')[0]} is invalid on surrogate`
-      );
-    });
-  }
-
-  it('warns if bindable element has more attributes other than the allowed', function () {
-    const template = `<template as-custom-element="foo-bar">
-      <bindable property="prop" unknown-attr who-cares="no one"></bindable>
-    </template>
-    <div></div>`;
-    const { container, sut } = createFixture();
-
-    sut.compile({ name: 'lorem-ipsum', template }, container, null);
-    if (__DEV__) {
-      const sinks = container.get(DefaultLogger)['warnSinks'] as ISink[];
-      const eventLog = sinks.find((s) => s instanceof EventLog) as EventLog;
-      assert.strictEqual(eventLog.log.length, 1, `eventLog.log.length`);
-      const event = eventLog.log[0];
-      assert.strictEqual(event.severity, LogLevel.warn);
-      assert.includes(
-        event.toString(),
-        'The attribute(s) unknown-attr, who-cares will be ignored for <bindable property="prop" unknown-attr="" who-cares="no one"></bindable>. Only property, attribute, mode are processed.'
-      );
-    }
-  });
-});
-
-describe('TemplateCompiler - au-slot', function () {
-  function createFixture() {
-    const ctx = TestContext.create();
-    const container = ctx.container;
-    const sut = ctx.templateCompiler;
-    return { ctx, container, sut };
-  }
-  function $createCustomElement(template: string, name: string) {
-    return CustomElement.define({ name, isStrictBinding: true, template, bindables: { people: { mode: BindingMode.default } }, }, class MyElement { });
-  }
-
-  class ExpectedSlotFallbackInfo {
-    public constructor(
-      public readonly slotName: string,
-      public readonly content: string,
-    ) { }
-  }
-  class TestData {
-    public constructor(
-      public readonly template: string,
-      public readonly customElements: CustomElementType[],
-      public readonly allExpectedProjections: [string, Record<string, string>][] | null,
-      public readonly expectedSlotInfos: ExpectedSlotFallbackInfo[],
-    ) {
-    }
-  }
-
-  function *getTestData() {
-    // projections verification
-    yield new TestData(
-      `<my-element><div au-slot></div></my-element>`,
-      [$createCustomElement('', 'my-element')],
-      [['my-element', { 'default': '<div></div>' }]],
-      [],
-    );
-    yield new TestData(
-      `<my-element><div au-slot="s1">p1</div><div au-slot="s2">p2</div></my-element>`,
-      [$createCustomElement('', 'my-element')],
-      [['my-element', { s1: '<div>p1</div>', s2: '<div>p2</div>' }]],
-      [],
-    );
-    yield new TestData(
-      `<my-element><div au-slot="s1">p1</div><div au-slot="s1">p2</div></my-element>`,
-      [$createCustomElement('', 'my-element')],
-      [['my-element', { s1: '<div>p1</div><div>p2</div>' }]],
-      [],
-    );
-    yield new TestData(
-      `<my-element><au-slot au-slot><div au-slot="s1">p1</div><div au-slot="s1">p2</div></au-slot></my-element>`,
-      [$createCustomElement('', 'my-element')],
-      [['my-element', { 'default': '<au-m class="au"></au-m>' }]],
-      [],
-    );
-
-    // fallback verification
-    yield new TestData(
-      `<au-slot name="s1">s1fb</au-slot><au-slot name="s2"><div>s2fb</div></au-slot>`,
-      [],
-      null,
-      [
-        new ExpectedSlotFallbackInfo('s1', 's1fb'),
-        new ExpectedSlotFallbackInfo('s2', '<div>s2fb</div>'),
-      ],
-    );
-    yield new TestData(
-      `<au-slot name="s1">s1fb</au-slot><au-slot name="s2"><div>s2fb</div></au-slot>`,
-      [],
-      null,
-      [
-        new ExpectedSlotFallbackInfo('s1', 's1fb'),
-        new ExpectedSlotFallbackInfo('s2', '<div>s2fb</div>'),
-      ],
-    );
-    yield new TestData(
-      `<au-slot name="s1">s1fb</au-slot><au-slot name="s2"><div>s2fb</div></au-slot>`,
-      [],
-      null,
-      [
-        new ExpectedSlotFallbackInfo('s1', 's1fb'),
-        new ExpectedSlotFallbackInfo('s2', '<div>s2fb</div>'),
-      ],
-    );
-    yield new TestData(
-      `<au-slot name="s1">s1fb</au-slot><my-element><div au-slot>p</div></my-element>`,
-      [$createCustomElement('', 'my-element')],
-      null,
-      [
-        new ExpectedSlotFallbackInfo('s1', 's1fb'),
-      ],
-    );
-  }
-  for (const { customElements, template, expectedSlotInfos, allExpectedProjections } of getTestData()) {
-    it(`compiles - ${template}`, function () {
-      const { sut, container } = createFixture();
-      container.register(AuSlot, ...customElements);
-
-      const compiledDefinition = sut.compile(
-        CustomElementDefinition.create({ name: 'my-ce', template }, class MyCe { }),
-        container,
-        { projections: null }
-      );
-
-      type HEI = HydrateElementInstruction;
-      const allInstructions = compiledDefinition.instructions.flat();
-      for (const expectedSlotInfo of expectedSlotInfos) {
-        const actualInstruction = allInstructions.find((i) =>
-          i.type === InstructionType.hydrateElement
-          && (typeof (i as HEI).res === 'string' && ((i as HEI).res as string).includes('au-slot')
-            || (i as HEI).res === CustomElement.getDefinition(AuSlot)
-          )
-          && (i as HydrateElementInstruction).auSlot.name === expectedSlotInfo.slotName
-        ) as HydrateElementInstruction;
-        assert.notEqual(actualInstruction, void 0, 'instruction');
-        const actualSlotInfo = actualInstruction.auSlot;
-        assert.deepStrictEqual((actualSlotInfo.fallback.template as HTMLElement).outerHTML, `<template>${expectedSlotInfo.content}</template>`, 'content');
-        assert.deepStrictEqual(actualSlotInfo.fallback.needsCompile, false, 'needsCompile');
-      }
-
-      // for each element instruction found
-      // verify projections for it compiles properly
-      if (allExpectedProjections == null) {
-        return;
-      }
-
-      for (const [elName, projections] of allExpectedProjections) {
-        const elementInstruction = allInstructions.find(i =>
-          i.type === InstructionType.hydrateElement
-          && (typeof (i as HEI).res === 'string' && ((i as HEI).res as string) === elName
-            || (i as HEI).res === container.find(CustomElement, elName)
-          )
-        ) as HydrateElementInstruction;
-        assert.notEqual(elementInstruction, void 0, `Instruction for element "${elName}" missing`);
-        const actualProjections = elementInstruction.projections;
-        for (const slotName in projections) {
-          const def = actualProjections[slotName];
-          assert.instanceOf(def, CustomElementDefinition);
-          assert.deepStrictEqual((def.template as HTMLElement).outerHTML, `<template>${projections[slotName]}</template>`, 'content');
-          assert.deepStrictEqual(def.needsCompile, false, 'needsCompile');
-        }
-      }
-    });
-  }
-});
-
-describe('TemplateCompiler - hooks', function () {
-  function createFixture() {
-    const ctx = TestContext.create();
-    const container = ctx.container;
-    const sut = ctx.templateCompiler;
-    return { ctx, container, sut };
-  }
-
-  it('invokes before compile hooks', function () {
-    const template = `<template></template>`;
-    const { container, sut } = createFixture();
-    let hookCallCount = 0;
-
-    container.register(Registration.instance(ITemplateCompilerHooks, {
-      compiling(template: HTMLElement) {
-        hookCallCount++;
-        template.setAttribute('data-hello', 'world');
-      }
-    }));
-
-    const definition = sut.compile({ name: 'lorem-ipsum', template }, container, null);
-    assert.strictEqual(hookCallCount, 1);
-    assert.strictEqual((definition.template as Element).getAttribute('data-hello'), 'world');
-  });
-
-  it('invokes all hooks', function () {
-    const template = `<template></template>`;
-    const { container, sut } = createFixture();
-    let hookCallCount = 0;
-
-    container.register(Registration.instance(ITemplateCompilerHooks, {
-      compiling(template: HTMLElement) {
-        hookCallCount++;
-        template.setAttribute('data-hello', 'world');
-      }
-    }));
-    container.register(Registration.instance(ITemplateCompilerHooks, {
-      compiling(template: HTMLElement) {
-        hookCallCount++;
-        template.setAttribute('data-world', 'hello');
-      }
-    }));
-
-    const definition = sut.compile({ name: 'lorem-ipsum', template }, container, null);
-    assert.strictEqual(hookCallCount, 2);
-    assert.strictEqual((definition.template as Element).getAttribute('data-hello'), 'world');
-    assert.strictEqual((definition.template as Element).getAttribute('data-world'), 'hello');
-  });
-
-  it('does not throw if the compile hooks does not have any hooks', function () {
-    const template = `<template></template>`;
-    const { container, sut } = createFixture();
-
-    container.register(Registration.instance(ITemplateCompilerHooks, {}));
-    assert.doesNotThrow(() => sut.compile({ name: 'lorem-ipsum', template }, container, null));
-  });
-
-  it('invokes hooks with resources semantic - only leaf', function () {
-    const template = `<template></template>`;
-    const { container, sut } = createFixture();
-    let hookCallCount = 0;
-    const createResolver = () => Registration.instance(ITemplateCompilerHooks, {
-      compiling(template: HTMLElement) {
-        hookCallCount++;
-        template.setAttribute('data-hello', 'world');
-      }
-    });
-    const middleContainer = container.createChild();
-    const leafContainer = middleContainer.createChild();
-    middleContainer.register(createResolver());
-    leafContainer.register(createResolver());
-
-    const definition = sut.compile({ name: 'lorem-ipsum', template }, leafContainer, null);
-    assert.strictEqual(hookCallCount, 1);
-    assert.strictEqual((definition.template as Element).getAttribute('data-hello'), 'world');
-  });
-
-  it('invokes hooks with resources semantic - leaf + root', function () {
-    const template = `<template></template>`;
-    const { container, sut } = createFixture();
-    let hookCallCount = 0;
-    const createResolver = (value: string) => Registration.instance(ITemplateCompilerHooks, {
-      compiling(template: HTMLElement) {
-        hookCallCount++;
-        template.setAttribute(`data-${value}`, value);
-      }
-    });
-    const middleContainer = container.createChild();
-    const leafContainer = middleContainer.createChild();
-    container.register(createResolver('root'));
-    middleContainer.register(createResolver('middle'));
-    leafContainer.register(createResolver('leaf'));
-
-    const definition = sut.compile({ name: 'lorem-ipsum', template }, leafContainer, null);
-    assert.strictEqual(hookCallCount, 2);
-    assert.strictEqual((definition.template as Element).getAttribute('data-root'), 'root');
-    assert.strictEqual((definition.template as Element).getAttribute('data-middle'), null);
-    assert.strictEqual((definition.template as Element).getAttribute('data-leaf'), 'leaf');
   });
 });
